requirements.txt
flask
flask-sqlalchemy
flask-script
wtforms
mailchimp
gevent